rename some files in nova-lib, dash to underline
